Track mathics-core API repl change for Dialog[]
