+++ .gitignore
@@ -35,6 +35,7 @@
 /ChangeLog.orig
 /ChangeLog.rej
 /ChangeLog.spell-corrected
+/Mathics3_Frontend_CLI.egg-info
 /PKG-INFO
 /README
 /README.html
